Extract RBI helper
This should be made available to all compilers.
