+++ lib/tapioca/internal.rb
@@ -8,6 +8,7 @@ require "tapioca/sorbet_ext/fixed_hash_patch"
 require "tapioca/generic_type_registry"
 require "tapioca/helpers/cli_helper"
 require "tapioca/helpers/config_helper"
+require "tapioca/helpers/rbi_helper"
 require "tapioca/helpers/shims_helper"
 require "tapioca/helpers/sorbet_helper"
 require "tapioca/generators"
